+++ internal/command/local/printers.go
@@ -8,6 +8,7 @@ import (
 	"github.com/signadot/cli/internal/config"
 	commonapi "github.com/signadot/cli/internal/locald/api"
 	sbmapi "github.com/signadot/cli/internal/locald/api/sandboxmanager"
+	sbmgr "github.com/signadot/cli/internal/locald/sandboxmanager"
 	connectcfg "github.com/signadot/libconnect/config"
 )
 
@@ -16,7 +17,7 @@ func printRawStatus(cfg *config.LocalStatus, out io.Writer, printer func(out io.
 	// unmarshal the ci config
 	ciConfig, err := sbmapi.ToCIConfig(status.CiConfig)
 	if err != nil {
-		return fmt.Errorf("couldn't unmarshal ci-config from sandbox manager status, %v", err)
+		return fmt.Errorf("couldn't unmarshal ci-config from sandboxmanager status, %v", err)
 	}
 
 	// convert the status into a map (useful to convert snake-case fields to camel-case,
@@ -237,37 +238,12 @@ func getRawPortforward(cfg *config.LocalStatus, ciConfig *config.ConnectInvocati
 	return result
 }
 
-func checkLocalStatusConnectErrors(ciConfig *config.ConnectInvocationConfig, status *sbmapi.StatusResponse) []error {
-	var errs []error
-	// check port forward status
-	if ciConfig.ConnectionConfig.Type == connectcfg.PortForwardLinkType {
-		err := checkPortforwardStatus(status.Portforward)
-		if err != nil {
-			errs = append(errs, err)
-		}
-	}
-	// check root manager (if running)
-	if ciConfig.WithRootManager {
-		// check localnet service
-		err := checkLocalNetStatus(status.Localnet)
-		if err != nil {
-			errs = append(errs, err)
-		}
-		// check hosts service
-		err = checkHostsStatus(status.Hosts)
-		if err != nil {
-			errs = append(errs, err)
-		}
-	}
-	return errs
-}
-
 func printLocalStatus(cfg *config.LocalStatus, out io.Writer, status *sbmapi.StatusResponse) error {
 	ciConfig, err := sbmapi.ToCIConfig(status.CiConfig)
 	if err != nil {
 		return fmt.Errorf("couldn't unmarshal ci-config from sandbox manager status, %v", err)
 	}
-	connectErrs := checkLocalStatusConnectErrors(ciConfig, status)
+	connectErrs := sbmgr.CheckStatusConnectErrors(status, ciConfig)
 
 	// create a printer
 	printer := statusPrinter{
@@ -288,51 +264,6 @@ func printLocalStatus(cfg *config.LocalStatus, out io.Writer, status *sbmapi.Sta
 		printer.printErrors(connectErrs)
 	}
 	return nil
-}
-
-func checkPortforwardStatus(portforward *commonapi.PortForwardStatus) error {
-	errorMsg := "failed to establish port-forward"
-	if portforward != nil {
-		if portforward.Health != nil {
-			if portforward.Health.Healthy {
-				return nil
-			}
-			if portforward.Health.LastErrorReason != "" {
-				errorMsg += fmt.Sprintf(" (%q)", portforward.Health.LastErrorReason)
-			}
-		}
-	}
-	return fmt.Errorf(errorMsg)
-}
-
-func checkLocalNetStatus(localnet *commonapi.LocalNetStatus) error {
-	errorMsg := "failed to setup localnet"
-	if localnet != nil {
-		if localnet.Health != nil {
-			if localnet.Health.Healthy {
-				return nil
-			}
-			if localnet.Health.LastErrorReason != "" {
-				errorMsg += fmt.Sprintf(" (%q)", localnet.Health.LastErrorReason)
-			}
-		}
-	}
-	return fmt.Errorf(errorMsg)
-}
-
-func checkHostsStatus(hosts *commonapi.HostsStatus) error {
-	errorMsg := "failed to configure hosts in /etc/hosts"
-	if hosts != nil {
-		if hosts.Health != nil {
-			if hosts.Health.Healthy {
-				return nil
-			}
-			if hosts.Health.LastErrorReason != "" {
-				errorMsg += fmt.Sprintf(" (%q)", hosts.Health.LastErrorReason)
-			}
-		}
-	}
-	return fmt.Errorf(errorMsg)
 }
 
 type statusPrinter struct {
