--- internal/locald/api/sandboxmanager/api_error.go
@@ -1,9 +1,0 @@
-package sandboxmanager
-
-func APIErrorResponse(err error) *ApplySandboxResponse {
-	return &ApplySandboxResponse{
-		It: &ApplySandboxResponse_ApiError{
-			ApiError: err.Error(),
-		},
-	}
-}
