+++ internal/locald/sandboxmanager/sandbox_monitor.go
@@ -6,11 +6,12 @@ import (
 	"sync"
 	"time"
 
+	"log/slog"
+
 	"github.com/signadot/go-sdk/models"
 	clapi "github.com/signadot/libconnect/apiv1"
 	clapiclient "github.com/signadot/libconnect/common/apiclient"
 	"github.com/signadot/libconnect/revtun"
-	"log/slog"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
